MMIO changes
* Change MMIO data to u32 instead of u64 to fix memoy bug
* Add some currently unused MMIO registers.

+++ src/peripherals/uart.rs
@@ -27,7 +27,7 @@ impl Uart {
         while mmio_read(mmio::UART_FR).bit(5) {
             hint::spin_loop();
         }
-        mmio_write(mmio::UART0_BASE, byte as u64);
+        mmio_write(mmio::UART0_BASE, byte as u32);
     }
 }
